+++ brozzler/cli.py
@@ -30,16 +30,16 @@ import doublethink
 import signal
 import string
 import structlog
-import subprocess
 import sys
 import threading
 import time
 import traceback
 import warnings
 import yaml
-import shutil
 import base64
 import rethinkdb as rdb
+
+from brozzler import suggest_default_chrome_exe
 
 r = rdb.RethinkDB()
 
@@ -211,68 +211,6 @@ def configure_logging(args):
     warnings.simplefilter(
         "ignore", category=requests.packages.urllib3.exceptions.InsecurePlatformWarning
     )
-
-
-def mdfind(identifier):
-    try:
-        result = subprocess.check_output(
-            ["mdfind", f"kMDItemCFBundleIdentifier == {identifier}"], text=True
-        )
-    # Just treat any errors as "couldn't find app"
-    except subprocess.CalledProcessError:
-        return None
-
-    if result:
-        return result.rstrip("\n")
-
-
-def suggest_default_chrome_exe_mac():
-    path = None
-    # Try Chromium first, then Chrome
-    result = mdfind("org.chromium.Chromium")
-    if result is not None:
-        path = f"{result}/Contents/MacOS/Chromium"
-
-    result = mdfind("com.google.Chrome")
-    if result is not None:
-        path = f"{result}/Contents/MacOS/Google Chrome"
-
-    if path is not None and os.path.exists(path):
-        return path
-
-    # Fall back to default paths if mdfind couldn't find it
-    # (mdfind might fail to find them even in their default paths
-    # if the system has Spotlight disabled.)
-    for path in [
-        "/Applications/Chromium.app/Contents/MacOS/Chromium",
-        "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome",
-    ]:
-        if os.path.exists(path):
-            return path
-
-
-def suggest_default_chrome_exe():
-    # First ask mdfind, which lets us find it in non-default paths
-    if sys.platform == "darwin":
-        path = suggest_default_chrome_exe_mac()
-        if path is not None:
-            return path
-
-    # "chromium-browser" is the executable on ubuntu trusty
-    # https://github.com/internetarchive/brozzler/pull/6/files uses "chromium"
-    # google chrome executable names taken from these packages:
-    # http://www.ubuntuupdates.org/ppa/google_chrome
-    for exe in [
-        "chromium-browser",
-        "chromium",
-        "google-chrome",
-        "google-chrome-stable",
-        "google-chrome-beta",
-        "google-chrome-unstable",
-    ]:
-        if shutil.which(exe):
-            return exe
-    return "chromium-browser"
 
 
 class BetterArgumentDefaultsHelpFormatter(argparse.ArgumentDefaultsHelpFormatter):
